add suggestViewingKey: boolean param to suggestToken api
suggestViewingKey defaults to true.

This is needed because all apps today expect viewing keys not permits. So if a user adds a token with permit manually, calling getSecret20ViewingKey will obviously return nothing. So the app needs to be able to request that the user overwrite the permit query authorization with a viewing key type authorization